validate URL function with type
