+++ jamming-app/apps/jamming-api/gloabal.d.ts
@@ -1,0 +1,6 @@
+import { StringSchema } from 'yup';
+declare module 'yup' {
+  interface StringSchema {
+    validateURL(url: string): this;
+  }
+}
