+++ jamming-app/apps/jamming-api/tsconfig.spec.json
@@ -9,6 +9,7 @@
     "jest.config.ts",
     "src/**/*.test.ts",
     "src/**/*.spec.ts",
-    "src/**/*.d.ts"
+    "src/**/*.d.ts",
+    "gloabal.d.ts"
   ]
 }
